requirements.txt
h5py==3.1.0
numpy==1.19.5
Pillow==8.4.0
torch==1.0.0
pandas~=1.1.5
scikit-image~=0.17.2
tqdm~=4.62.3
scipy~=1.2.1
natsort~=8.0.0
warmup_scheduler==0.3